fas_simple_distill/requirements.txt
opencv-contrib-python==4.5.2.54
git+https://github.com/tmbdev/webdataset.git@c22bbcd11c0e4a12418e56107687b76ff07cfd92
addict==2.4.0
iglovikov-helper-functions==0.0.53
mlflow>=1.17.0
git+https://github.com/rwightman/pytorch-image-models
timm>=0.5.4
pytorch-metric-learning==0.9.99
pyyaml
torchinfo
torch_pruning
git+https://github.com/Lightning-AI/metrics.git@release/stable
ray[tune]
hyperopt
kornia
termcolor
imgaug
onnxruntime
petyr
pillow
pytorch-msssim
torchdata==0.3
mediapipe